utils.py: Fix Matplotlib warning:
  UserWarning: color is redundantly defined by the 'color' keyword
  argument and the fmt string "yv" (-> color='y'). The keyword
  argument will take precedence.

This warning started due to a Matplotlib upgrade. The format string
"yv" has been in NEMO for a very long time and, as far as I can tell,
was incorrect from the beginning. The color keyword was correctly
forcing the markers to be red, so this error was never evident.